fix: revert Decimal changes from previous commit

+++ packages/andromeda-math/src/curve.rs
@@ -1,6 +1,6 @@
 use andromeda_std::{amp::AndrAddr, andr_exec, andr_instantiate, andr_query, error::ContractError};
 use cosmwasm_schema::{cw_serde, QueryResponses};
-use cosmwasm_std::{ensure, Decimal};
+use cosmwasm_std::ensure;
 
 #[andr_instantiate]
 #[cw_serde]
@@ -60,7 +60,7 @@ pub enum QueryMsg {
     #[returns(GetCurveConfigResponse)]
     GetCurveConfig {},
     #[returns(GetPlotYFromXResponse)]
-    GetPlotYFromX { x_value: Decimal },
+    GetPlotYFromX { x_value: u64 },
 }
 
 #[cw_serde]
